Custom error validation and checks added

+++ UserRegistrationSystemRefresh/src/main/java/com/boyan/Rest/UserRegistrationRestController.java
@@ -1,6 +1,7 @@
 package com.boyan.Rest;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.boyan.Exception.CustomErrorType;
 import com.boyan.dto.UserDTO;
 import com.boyan.repository.UserJpaRepository;
 
@@ -24,60 +26,84 @@ import com.boyan.repository.UserJpaRepository;
 @RequestMapping(value = "/api/user")
 public class UserRegistrationRestController {
 
-	public static final Logger logger = LoggerFactory
-			.getLogger(UserRegistrationRestController.class);
-	
-	
+	public static final Logger logger = LoggerFactory.getLogger(UserRegistrationRestController.class);
+
 	private UserJpaRepository repository;
-	
+
 	@Autowired
 	public UserRegistrationRestController(UserJpaRepository repository) {
 		this.repository = repository;
 	}
-	
+
 	@GetMapping(value = "/")
 	public ResponseEntity<List<UserDTO>> listUsers() {
 		List<UserDTO> users = repository.findAll();
-		
+		if (users.isEmpty()) {
+			return new ResponseEntity<List<UserDTO>>(HttpStatus.NO_CONTENT);
+		}
+
 		return new ResponseEntity<List<UserDTO>>(users, HttpStatus.OK);
 	}
-	
+
 	@PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<UserDTO> createUser(@RequestBody final UserDTO user) {
+		if (repository.findByName(user.getName()) != null) {
+			return new ResponseEntity<UserDTO>(
+					new CustomErrorType("User with the same " + user.getName() + " already exist"),
+					HttpStatus.CONFLICT);
+		}
+
 		repository.save(user);
-		return new ResponseEntity<UserDTO>(repository.findByName(user.getName()), 
-				HttpStatus.CREATED);
+		return new ResponseEntity<UserDTO>(repository.findByName(user.getName()), HttpStatus.CREATED);
 	}
-	
+
 	@GetMapping(value = "/{id}")
 	public ResponseEntity<UserDTO> getUserById(@PathVariable("id") final Long id) {
-		return new ResponseEntity<UserDTO>(repository.findById(id).get(),
-				HttpStatus.OK);
+		Optional<UserDTO> user = repository.findById(id);
+
+		if (user.isEmpty()) {
+			return new ResponseEntity<UserDTO>(new CustomErrorType("User with id " + id + " not found"),
+					HttpStatus.NOT_FOUND);
+		}
+
+		return new ResponseEntity<UserDTO>(user.get(), HttpStatus.OK);
+
 	}
-	
+
 	@PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<UserDTO> updateUser(@PathVariable("id") final Long id,
-			@RequestBody UserDTO user) {
+	public ResponseEntity<UserDTO> updateUser(@PathVariable("id") final Long id, @RequestBody UserDTO user) {
 		// fetch user based on id and set it to currentUser object of type UserDTO
-		UserDTO currentUser = repository.findById(id).get();
-		
+		UserDTO currentUser = repository.findById(id).orElse(null);
+
+		if (currentUser == null) {
+			return new ResponseEntity<UserDTO>(
+					new CustomErrorType("Unable to update." + "User with id " + id + " not found"),
+					HttpStatus.NOT_FOUND);
+		}
+
 		// update currentUser object data with user object data
 		currentUser.setName(user.getName());
 		currentUser.setAddress(user.getAddress());
 		currentUser.setEmail(user.getEmail());
-		
+
 		// save current user object
 		repository.saveAndFlush(currentUser);
-		
-		return new ResponseEntity<UserDTO>(repository.findById(id).get(),
-				HttpStatus.OK);
+
+		return new ResponseEntity<UserDTO>(repository.findById(id).get(), HttpStatus.OK);
 	}
-	
+
 	@DeleteMapping(value = "/{id}")
 	public ResponseEntity<UserDTO> deleteUser(@PathVariable("id") final Long id) {
-		repository.deleteById(id);
+		Optional<UserDTO> userForDeletion = repository.findById(id);
 		
+		if (userForDeletion.isEmpty()) {
+			return new ResponseEntity<UserDTO>(new CustomErrorType("Unable to delete. "
+					+ "User with id " + id + " not found."), HttpStatus.NOT_FOUND);
+		}
+		
+		repository.deleteById(id);
+
 		return new ResponseEntity<UserDTO>(HttpStatus.NO_CONTENT);
 	}
-	
+
 }
